pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-pulse"
version = "0.1.0"
description = "Agent autonome de veille IA — briefing quotidien via LangGraph"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

# -------------------------------------------------------------------
# Ruff
# -------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "B008", "SIM108"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "N806", "F841", "RUF059"]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# -------------------------------------------------------------------
# Mypy
# -------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["langgraph.*", "langchain.*", "langchain_core.*", "langchain_anthropic.*", "feedparser.*", "bs4.*", "yaml.*"]
ignore_missing_imports = true
ignore_errors = true

# -------------------------------------------------------------------
# Pytest
# -------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
]

# -------------------------------------------------------------------
# Coverage
# -------------------------------------------------------------------
[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
